:hammer: First Version (WIP)
Still a bunch to do like building and sending the embed automatically, but the API call and auth to twitch i got knocked out tonight which is a good start. As well as learning discord py for the first time this was a decently spent hour and a half. Also got discord and twitch bot integration set up for this.